requirements.txt
python-dotenv
requests
appscript
osascript
faker
emoji
